@@ -1,26 +1,25 @@
 {
-  "name": "wvglobe",
+  "name": "globetaketwo",
   "version": "0.1.0",
   "private": true,
-  "proxy": "http://localhost:3001",
   "dependencies": {
-    "@react-three/drei": "^7.0.2",
-    "@react-three/fiber": "^7.0.1",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
+    "@apollo/client": "^3.6.9",
+    "@react-three/drei": "^9.28.2",
+    "@react-three/fiber": "^8.7.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "bootstrap": "^5.2.1",
+    "graphql": "^16.6.0",
+    "jwt-decode": "^3.1.2",
+    "react": "^18.2.0",
+    "react-bootstrap": "^1.0.1",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
     "reactdom": "^2.0.0",
-    "react-router-dom": "^6.2.1",
-    "react-bootstrap": "1.0.1",
-    "styled-components": "^5.3.0",
-    "three": "^0.129.0",
-    "@apollo/client": "^3.5.8",
-    "bootstrap": "^4.4.1",
-    "graphql": "^15.4.0",
-    "jwt-decode": "^2.2.0"
+    "styled-components": "^5.3.5",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
