@@ -8,20 +8,21 @@ import {
 } from '@apollo/client';
 import { setContext } from '@apollo/client/link/context';
 
-import './App.css';
+// import './App.css';
+import './index.css';
 import styled from 'styled-components';
 import { Canvas } from '@react-three/fiber';
 import { Suspense } from 'react';
 import { Earth } from './components/earth';
-// import { TopSection } from "./components/topSection";
 
 import SearchBooks from './pages/SearchBooks';
 import SavedBooks from './pages/SavedBooks';
 import Navbar from './components/Navbar';
+import topSection from './components/topSection/index.js'
 
 const CanvasContainer = styled.div`
-  width: 60%;
-  height: 60%;
+  width: 100%;
+  height: 100%;
 `;
 
 // Construct our main GraphQL API endpoint
@@ -66,7 +67,6 @@ function App() {
           </>
         </Router>
       </ApolloProvider>
-      {/* <TopSection /> */}
       <Canvas>
         <Suspense fallback={null}>
           <Earth />
