@@ -1,5 +1,6 @@
 import React, { useRef } from 'react';
-import { useFrame, useLoader } from '@react-three/fiber';
+import { useLoader, useFrame } from '@react-three/fiber';
+import { TextureLoader } from 'three';
 import { OrbitControls, Stars } from '@react-three/drei';
 import * as THREE from 'three';
 
@@ -7,7 +8,7 @@ import EarthDayMap from '../../assets/textures/8k_earth_daymap.jpg';
 import EarthNormalMap from '../../assets/textures/8k_earth_normal_map.jpg';
 import EarthSpecularMap from '../../assets/textures/8k_earth_specular_map.jpg';
 import EarthCloudsMap from '../../assets/textures/8k_earth_clouds.jpg';
-import { TextureLoader } from 'three';
+import EarthNightMap from '../../assets/textures/8k_earth_nightmap.jpg';
 
 export function Earth(props) {
   const [colorMap, normalMap, specularMap, cloudsMap] = useLoader(
@@ -15,20 +16,20 @@ export function Earth(props) {
     [EarthDayMap, EarthNormalMap, EarthSpecularMap, EarthCloudsMap]
   );
 
-  const earthRef = useRef();
-  const cloudsRef = useRef();
+  // const earthRef = useRef();
+  // const cloudsRef = useRef();
 
   // useFrame(({ clock }) => {
   //   const elapsedTime = clock.getElapsedTime();
 
-  //   earthRef.current.rotation.y = elapsedTime / 9;
-  //   cloudsRef.current.rotation.y = elapsedTime / 9;
+  //   earthRef.current.rotation.y = elapsedTime / 6;
+  //   cloudsRef.current.rotation.y = elapsedTime / 6;
   // });
 
   return (
     <>
-      {/* <ambientLight intensity={1} /> */}
-      <pointLight color="#f6f3ea" position={[2, 0, 5]} intensity={1.2} />
+      <ambientLight intensity={1} />
+      {/* <pointLight color="#f6f3ea" position={[2, 0, 2]} intensity={1.2} /> */}
       <Stars
         radius={300}
         depth={60}
@@ -37,7 +38,7 @@ export function Earth(props) {
         saturation={0}
         fade={true}
       />
-      {/* <mesh ref={cloudsRef} position={[0, 0, 3]}> */}
+      {/* <mesh ref={cloudsRef}> */}
       <mesh>
         <sphereGeometry args={[1.005, 32, 32]} />
         <meshPhongMaterial
@@ -48,15 +49,14 @@ export function Earth(props) {
           side={THREE.DoubleSide}
         />
       </mesh>
-      {/* <mesh ref={earthRef} position={[0, 0, 3]}> */}
+      {/* <mesh ref={earthRef}> */}
       <mesh>
         <sphereGeometry args={[1, 32, 32]} />
         <meshPhongMaterial specularMap={specularMap} />
         <meshStandardMaterial
           map={colorMap}
           normalMap={normalMap}
-          metalness={0.4}
-          roughness={0.7}
+          metalness={0.7}
         />
         <OrbitControls
           enableZoom={true}
@@ -64,7 +64,7 @@ export function Earth(props) {
           enableRotate={true}
           zoomSpeed={0.6}
           panSpeed={0.5}
-          rotateSpeed={0.4}
+          rotationSpeed={0.4}
         />
       </mesh>
     </>
