@@ -1,7 +1,7 @@
-const mongoose = require('mongoose');
+const mongoose = require("mongoose");
 
 mongoose.connect(
-  process.env.MONGODB_URI || 'mongodb://localhost:27017/googlebooks',
+  process.env.MONGODB_URI || "mongodb://localhost:27017/worldview",
   {
     useNewUrlParser: true,
     useUnifiedTopology: true,
